Created mad2 method for array class
mad2 was not working correctly when byrow=TRUE.  In my case mad was a list of vectors and it could not assign rownames to an object with no dimension.  So I added a clause to not assign rownames when dim(mad[[1]]) is NULL.  I am not sure if it works when there is more than one chromosome.  Rob should check this out.