Included rowMAD in export statement in NAMESPACE.
I was concerned about Rob's caveat about %dopar% with worker nodes, so I added rowMAD to the list of exported functions in MinimumDistance.  In the file mad-methods we have VanillaICE:::rowMAD and I wanted to treat this very carefully.  I have not removed VanillaICE::: from this file --- as I want to clear it with Rob first.